+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## v3.2.1
 * PR 23: Added `Withdrawals.GetFeeEstimate`. Thanks la4atld!
 * PR 12: Added `Fees.GetCurrentFeesAsync` for maker and taker fees. Thanks vflame!
+* Issue 24: Added `L2UpdateEvent.Time` property to C# model for websocket events.
 
 ## v3.2.0
 * Added `ConnectResult` return type from `CoinbaseProWebSocket.ConnectAsync()` for better semantic connection handling.